created format_config.py file to hold classes used for Format creation and definition
