+++ tests/python_tests/helpers/stimuli_generator.py
@@ -2,7 +2,8 @@
 # SPDX-License-Identifier: Apache-2.0
 
 import torch
-from .format_arg_mapping import format_dict, DataFormat
+from .format_arg_mapping import format_dict
+from .format_config import DataFormat
 
 
 def flatten_list(sublists):
